build: migrate to ESM+CJS hybrid with separate package.json files
1. The dist folder has been split into ./esm/ and ./cjs/ sub-folders
2. Both of those sub-folders have a partial package.json file that do
not duplicate the full contents of the root package.json but instead just
declare one as an ESM module and the other one as a common JS module.
3. The resulting package passes the full checklist of hybrid packages
that verifies that typings and code are in the correct format.
4. A bash script was added to perform a post-build action that generates
the partial package.json files with the bare minimum information in them.
5. The typescript compiler configuration file was also split in two because
the only way (that I know) to implement a hybrid package is to run the
Typescript build twice, once for ESM and another time for CJS.

Signed-off-by: Peter Somogyvari <peter.metz@unarin.com>